+++ core/internal/api/api.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"github.com/hashicorp/go-retryablehttp"
+	"github.com/wandb/wandb/core/internal/auth"
 	"github.com/wandb/wandb/core/internal/clients"
 )
 
@@ -48,7 +49,7 @@ type Backend struct {
 	logger *slog.Logger
 
 	// Credentials to apply for backend requests.
-	credentialProvider CredentialProvider
+	credentialProvider auth.CredentialProvider
 }
 
 // An HTTP client for interacting with the W&B backend.
@@ -142,7 +143,7 @@ type BackendOptions struct {
 	Logger *slog.Logger
 
 	// Credentials to apply on every request.
-	CredentialProvider CredentialProvider
+	CredentialProvider auth.CredentialProvider
 }
 
 // Creates a [Backend].